+++ packages/vexide-panic/build.rs
@@ -1,6 +1,0 @@
-#![allow(missing_docs)]
-
-fn main() {
-    let manifest_dir = env!("CARGO_MANIFEST_DIR");
-    println!("cargo:rustc-link-search=native={manifest_dir}/link");
-}