+++ packages/vexide-startup/examples/booting.rs
@@ -8,23 +8,24 @@ use alloc::boxed::Box;
 
 use vex_sdk::vexTasksRun;
 use vexide_core::println;
-use vexide_startup::{CodeSignature, ProgramFlags, ProgramOwner, ProgramType};
+use vexide_startup::{
+    banner::themes::THEME_DEFAULT, CodeSignature, ProgramFlags, ProgramOwner, ProgramType,
+};
 
 #[no_mangle]
-extern "Rust" fn main() {
+unsafe extern "C" fn _start() -> ! {
     unsafe {
+        vexide_startup::startup::<true>(THEME_DEFAULT);
+
         // Write something to the screen to test if the program is running
         let test_box = Box::new(100);
         vex_sdk::vexDisplayRectFill(0, 0, *test_box, 200);
         println!("Hello, world!");
         vexTasksRun(); // Flush serial
     }
-}
 
-#[no_mangle]
-#[link_section = ".boot"]
-unsafe extern "C" fn _start() {
-    unsafe { vexide_startup::program_entry::<true>() }
+    // Exit once we're done.
+    vexide_core::program::exit();
 }
 
 #[link_section = ".code_signature"]