+++ packages/vexide-startup/src/lib.rs
@@ -1,21 +1,23 @@
-//! This crate provides a working entrypoint for the VEX V5 Brain.
+//! This crate provides a minimal startup routine for user code on the VEX V5 Brain.
 //!
-//! # Usage
+//! - User code begins at an assembly routine called `_boot`, which sets up the stack
+//!   section before jumping to a user-provided `_start` symbol, which should be your
+//!   rust entrypointt.
 //!
-//! Your entrypoint function should be an async function that takes a single argument of type [`Peripherals`](vexide_devices::peripherals::Peripherals).
-//! It can return any type implementing [`Termination`](vexide_core::program::Termination).
-//! ```rust
-//! #[vexide::main]
-//! async fn main(peripherals: Peripherals) { ... }
-//! ```
+//! - From there, the Rust entrypoint may call the [`startup`] function to finish the
+//!   startup process by clearing the `.bss` section (intended for uninitialized data)
+//!   and initializing vexide's heap allocator.
+//!
+//! This crate is NOT a crt0 implementation. No global constructors are called.
 
 #![no_std]
 #![feature(asm_experimental_arch)]
 #![allow(clippy::needless_doctest_main)]
 
+use banner::themes::BannerTheme;
 use bitflags::bitflags;
 
-mod banner;
+pub mod banner;
 
 /// Identifies the type of binary to VEXos.
 #[repr(u32)]
@@ -86,72 +88,78 @@ extern "C" {
     static mut __bss_end: u32;
 }
 
-extern "Rust" {
-    fn main();
-}
+// This is the true entrypoint of vexide, containing the first two
+// instructions of user code executed before anything else.
+//
+// This function loads the stack pointer to the stack region specified
+// in our linkerscript, then immediately branches to the Rust entrypoint
+// created by our macro.
+core::arch::global_asm!(
+    r#"
+.section .boot, "ax"
+.global _boot
 
-/// Sets up the user stack, zeroes the BSS section, and calls the user code.
-/// This function is designed to be used as an entry point for programs on the VEX V5 Brain.
+_boot:
+    ldr sp, =__stack_top @ Load the user stack.
+    b _start             @ Jump to the Rust entrypoint.
+"#
+);
+
+/// Zeroes the `.bss` section
 ///
-/// # Const Parameters
+/// # Arguments
 ///
-/// - `BANNER`: Enables the vexide startup banner, which prints the vexide logo ASCII art and a startup message.
+/// - `sbss`. Pointer to the start of the `.bss` section.
+/// - `ebss`. Pointer to the open/non-inclusive end of the `.bss` section.
+///   (The value behind this pointer will not be modified)
+/// - Use `T` to indicate the alignment of the `.bss` section.
 ///
 /// # Safety
 ///
-/// This function MUST only be called once and should only be called at the very start of program initialization.
-/// Calling this function more than one time will seriously mess up both your stack and your heap.
-pub unsafe fn program_entry<const BANNER: bool>() {
+/// - Must be called exactly once
+/// - `mem::size_of::<T>()` must be non-zero
+/// - `ebss >= sbss`
+/// - `sbss` and `ebss` must be `T` aligned.
+#[inline]
+unsafe fn zero_bss<T>(mut sbss: *mut T, ebss: *mut T)
+where
+    T: Copy,
+{
+    while sbss < ebss {
+        // NOTE(volatile) to prevent this from being transformed into `memclr`
+        unsafe {
+            core::ptr::write_volatile(sbss, core::mem::zeroed());
+            sbss = sbss.offset(1);
+        }
+    }
+}
+
+/// Startup Routine
+///
+/// - Sets up the heap allocator if necessary.
+/// - Zeroes the `.bss`` section if necessary.
+/// - Prints the startup banner with a specified theme, if enabled.
+///
+/// # Safety
+///
+/// Must be called once at the start of program execution after the stack has been setup.
+#[inline]
+pub unsafe fn startup<const BANNER: bool>(theme: BannerTheme) {
     #[cfg(target_arch = "arm")]
     unsafe {
-        use core::arch::asm;
-        asm!(
-            "
-            // Load the user stack
-            ldr sp, =__stack_top
-            "
+        // Fill the `.bss` section of our program's memory with zeroes to ensure that uninitialized data is allocated properly.
+        zero_bss(
+            core::ptr::addr_of_mut!(__bss_start),
+            core::ptr::addr_of_mut!(__bss_end),
         );
-    }
 
-    // Clear the BSS section (used for storing uninitialized data).
-    //
-    // VEXos doesn't do this for us on program start, so it's necessary here.
-    #[cfg(target_arch = "arm")]
-    unsafe {
-        use core::ptr::addr_of_mut;
-        let mut bss_start = addr_of_mut!(__bss_start);
-        while bss_start < addr_of_mut!(__bss_end) {
-            core::ptr::write_volatile(bss_start, 0);
-            bss_start = bss_start.offset(1);
-        }
-    }
-
-    unsafe {
         // Initialize the heap allocator
-        // This `cfg` is mostly just to make the language server happy. All of this code is near impossible to run in the WASM sim.
-        #[cfg(target_arch = "arm")]
+        // This cfg is mostly just to make the language server happy. All of this code is near impossible to run in the WASM sim.
         vexide_core::allocator::vexos::init_heap();
-        // Print the banner
-        if BANNER {
-            banner::print();
-        }
-        // Run VEXos background processing at a regular 2ms interval.
-        // This is necessary for serial and device reads to work properly.
-        vexide_async::task::spawn(async {
-            loop {
-                vex_sdk::vexTasksRun();
+    }
 
-                // In VEXCode programs, this is run in a tight loop with no delays, since they
-                // don't need to worry about running two schedulers on top of each other, but
-                // doing this in our case would cause this task to hog all the CPU time, which
-                // wouldn't allow futures to be polled in the async runtime.
-                vexide_async::time::sleep(::core::time::Duration::from_millis(2)).await;
-            }
-        })
-        .detach();
-        // Call the user code
-        main();
-        // Exit the program
-        vexide_core::program::exit();
+    // Print the banner
+    if BANNER {
+        banner::print(theme);
     }
 }