+++ CHANGELOG.md
@@ -24,12 +24,14 @@ Before releasing:
 ### Added
 
 - Added support for the V5 GPS Sensor (#79)
+- Added support for custom banner themes configurable through the `vexide::main` macro (#127)
 
 ### Fixed
 
 - Fixed an issue where the distance sensor relative_size returned a u32 when it can be negative. (#116)
 - Fixed an issue preventing the `Screen::draw_buffer` function from working properly. (#128)
 - Fixed an issue where panic messages would not be displayed even when the `display_panics` feature was enabled if the screens render mode was set to `DoubleBuffered`. (#134)
+- `GpsImu` should now validate on the correct port. (#141)
 
 ### Changed
 
@@ -42,7 +44,9 @@ Before releasing:
 - `SmartPort::device_type` now no longer returns a `Result`. (#121) (**Breaking Change**)
 - Updated the names of certain misspelled `enum` variants, constants, and fields. (#132) (**Breaking Change**)
 - Marks many futures as `#[must_use]` to warn when futures are created without `await`ing them. (#112)
-- Controller joystick axis getters now return `f64` instead of `f32`. (#133)
+- Changes the banner attribute syntax in the `vexide::main` macro. (#127) (**Breaking Change**)
+- Controller joystick axis getters now return `f64` instead of `f32`. (#133) (**Breaking Change**)
+- Fixed an issue where the async executor would block indefinetly on the first program run after a Brain reboot (#139)
 
 ### Removed